chore: extract PageContent function to another file, adjust ModeRules for other modes
